move core logic to lib/ directory
